% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_routines.R
\name{ac}
\alias{ac}
\title{adjust colors, while keeping the vector names}
\usage{
ac(x, alpha = 1, ...)
}
\arguments{
\item{x}{color vector}

\item{alpha}{transparenscy value (passed to adjustcolors as alpha.f)}

\item{...}{parameters passsed to adjustcolor}
}
\description{
adjust colors, while keeping the vector names
}
