% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_routines.R
\name{filter.genes.by.cluster.expression}
\alias{filter.genes.by.cluster.expression}
\title{Filter genes by requirining minimum average expression within at least one of the provided cell clusters}
\usage{
filter.genes.by.cluster.expression(emat, clusters,
  min.max.cluster.average = 0.1)
}
\arguments{
\item{emat}{spliced (exonic) count matrix}

\item{clusters}{named cell factor defining clusters}

\item{min.max.cluster.average}{required minimum average expression count (no normalization is perfomed)}
}
\value{
filtered emat matrix
}
\description{
Filter genes by requirining minimum average expression within at least one of the provided cell clusters
}
