% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_routines.R
\name{find.ip.sites}
\alias{find.ip.sites}
\title{identify positions of likely internal priming sites by looking for polyA/polyT stretches within
annotated intronic regions}
\usage{
find.ip.sites(gtf.file, genome, genome.name, w = 0.9, n = 15,
  min.score = "80\%", add.chr = TRUE)
}
\arguments{
\item{gtf.file}{location of a gtf file with gene annotations}

\item{genome}{bioC genome structure (i.e. Mmusculus)}

\item{genome.name}{name of the genome assembly (i.e. mm10)}

\item{w}{A/T weight in the PWM}

\item{n}{length of the motif}

\item{min.score}{minimal required match score}

\item{add.chr}{whether to add 'chr' prefix to the chromosome names in the gtf annotation (to match bioC)}
}
\value{
a data frame containing list of likely internal priming sites, listing chromosome ($chr), positions ($start/$end), name, PWM matching score, PWM match strand ($strand), gene, gene strand ($gs), and whether the motif is in concordant direction of gene transcription ($conc)
}
\description{
identify positions of likely internal priming sites by looking for polyA/polyT stretches within
annotated intronic regions
}
\examples{
\dontrun{
library(BSgenome.Mmusculus.UCSC.mm10)
ip.mm10 <- find.ip.sites('refdata-cellranger-mm10-1.2.0/genes/genes.gtf',Mmusculus,'mm10')
}
}
