% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_routines.R
\name{gene.relative.velocity.estimates}
\alias{gene.relative.velocity.estimates}
\title{Estimate RNA velocity using gene-relative slopes}
\usage{
gene.relative.velocity.estimates(emat, nmat, deltaT = 1, smat = NULL,
  steady.state.cells = colnames(emat), kCells = 10, cellKNN = NULL,
  kGenes = 1, old.fit = NULL, mult = 1000,
  min.nmat.smat.correlation = 0.05, min.nmat.emat.correlation = 0.05,
  min.nmat.emat.slope = 0.05, zero.offset = FALSE, deltaT2 = 1,
  fit.quantile = NULL, show.gene = NULL, do.par = TRUE,
  cell.dist = NULL, cell.emb = NULL, cell.colors = NULL,
  expression.gradient = NULL, residual.gradient = NULL,
  n.cores = defaultNCores(), verbose = TRUE)
}
\arguments{
\item{emat}{- spliced (exonic) count matrix}

\item{nmat}{- unspliced (nascent) count matrix}

\item{deltaT}{- amount of time to project the cell forward}

\item{smat}{- optional spanning read matrix (used in offset calculations)}

\item{steady.state.cells}{- optional set of steady-state cells on which the gamma should be estimated (defaults to all cells)}

\item{kCells}{- number of k nearest neighbors (NN) to use in slope calculation smoothing}

\item{cellKNN}{- optional pre-calculated cell KNN matrix}

\item{kGenes}{- number of genes (k) to use in gene kNN pooling}

\item{old.fit}{- optional old result (in this case the slopes and offsets won't be recalculated)}

\item{mult}{- library scaling factor (1e6 in case of FPM)}

\item{min.nmat.smat.correlation}{- minimum required Spearman rank correlation between n and s counts of a gene}

\item{min.nmat.emat.correlation}{- minimum required Spearman rank correlation between n and e counts of a gene}

\item{min.nmat.emat.slope}{- minimum sloope of n~e regression}

\item{zero.offset}{- should offset be set to zero, or determined (through smat regression or using near-0 e cases)}

\item{deltaT2}{- scaling of the projected difference vector (normally should be set to 1)}

\item{fit.quantile}{perform gamma fit on a top/bottom quantiles of expression magnitudes}

\item{show.gene}{an optional name of a gene for which the velocity estimation details should be shown (instead of estimating all velocities)}

\item{do.par}{whether the graphical device parameters should be reset as part of show.gene (default=TRUE)}

\item{cell.dist}{- cell distance to use in cell kNN pooling calculations}

\item{cell.emb}{- cell embedding to be used in show.gene function}

\item{cell.colors}{- cell colors to be used in show.gene function}

\item{expression.gradient}{- color palette used to show the expression magnitudes in show.gene function}

\item{residual.gradient}{- color palette used to show the u residuals in show.gene function}

\item{n.cores}{- number of cores to use}

\item{verbose}{- output messages about progress}
}
\value{
a list with velocity results, including the current normalized expression state ($current), projected ($projected), unscaled transcriptional change ($deltaE), fit results ($ko, $sfit), optional cell pooling parameters ($cellKNN, $kCells), kNN-convolved normalized matrices (conv.nmat.norm and conv.emat.norm)
}
\description{
Estimate RNA velocity using gene-relative slopes
}
