% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_routines.R
\name{global.velcoity.estimates}
\alias{global.velcoity.estimates}
\title{Structure-based gene velocity estimation}
\usage{
global.velcoity.estimates(emat, nmat, vel, base.df, deltaT = 1, smat = NULL,
  kGenes = 15, kGenes.trim = 5, smooth.kGenes = 0, kCells = 10,
  deltaT2 = 1, min.gene.conuts = 100, min.gene.cells = 20,
  min.intron.length = 10^3.5, min.exon.length = 10^2.7,
  top.global.pearson.deviance = 3, cellKNN = NULL, cell.dist = NULL,
  fit.quantile = NULL, m.pcount = 5, n.cores = defaultNCores())
}
\arguments{
\item{emat}{- spliced (exonic) count matrix}

\item{nmat}{- unspliced (nascent) count matrix}

\item{vel}{- initial gene-relative velocity estimates (output of the gene.relative.velocity.estimates function)}

\item{base.df}{gene structure information data frame ($base.df in output of read.smartseq2.bams), containing the following columns ($il - total intronic length in log10(length+1) scale; $el - total exonic length; $nex - number of expressed (above some low threshold) exons)}

\item{deltaT}{- amount of time to project the cell forward}

\item{smat}{- optional spanning read matrix (used in offset calculations)}

\item{kGenes}{- number of genes to use in evaluating trimmed mean of M values}

\item{kGenes.trim}{- number of genes to trim (from both ends)}

\item{smooth.kGenes}{- gene kNN pooling k value (used in the initial gene-relative fit)}

\item{kCells}{- number of k nearest neighbors (NN) to use in slope calculation smoothing}

\item{deltaT2}{- scaling of the projected difference vector (normally should be set to 1)}

\item{min.gene.conuts}{- minimum number of spliced reads/molecules that a gene should have}

\item{min.gene.cells}{- minimum number of cells in which a gene should be expressed}

\item{min.intron.length}{- minimum exon length}

\item{min.exon.length}{- minimum exon length}

\item{top.global.pearson.deviance}{- maximum deviance threshold to filter out genes with very high unsplied counts (likely due to other processes)}

\item{cellKNN}{- optional pre-calculated cell KNN matrix}

\item{cell.dist}{- cell distance to use in cell kNN pooling calculations}

\item{fit.quantile}{perform gamma fit on a top/bottom quantiles of expression magnitudes}

\item{m.pcount}{- pseudocount to be used in M value calculations (defaults to 5)}

\item{n.cores}{- number of cores to use}
}
\value{
a list with velocity results, including the current normalized expression state ($current), projected ($projected), unscaled transcriptional change ($deltaE), fit results ($ko, $sfit), optional cell pooling parameters ($cellKNN, $kCells), kNN-convolved normalized matrices (conv.nmat.norm and conv.emat.norm)
}
\description{
Structure-based gene velocity estimation
}
