% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_routines.R
\name{read.smartseq2.bams}
\alias{read.smartseq2.bams}
\title{read.smartseq2.bams}
\usage{
read.smartseq2.bams(bam.files, annotation.file, min.exon.count = 100,
  n.cores = defaultNCores())
}
\arguments{
\item{bam.files}{list of bam files}

\item{annotation.file}{refFlat genome annotation file (use gtfToGenePred to generate refFlat file from gtf)}

\item{min.exon.count}{minimum number of reads (across all cells) for an exon to be considered expressed in the dataset}

\item{n.cores}{number of cores to use}
}
\value{
a list containing: emat - exonic (spliced) read count matrix ; iomat - intronic (unspliced) matrix; smat - spanning read matrix; base.df - data frame containing gene structural information; exons - exon annotation and read counts; genes - gene annotation table with additional structural info; expr.lstat - gene length statistics when considering only expressed exons
}
\description{
DEPRECATED: Read in cell-specific bam files for SMART-seq2 measurement
This function is deprecated. Please use velocyto.py to prepare loom file from SMART-seq2 bam files.
}
