% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_routines.R
\name{read.strtc1.bams}
\alias{read.strtc1.bams}
\title{Read in cell-specific bam files for STRT/C1}
\usage{
read.strtc1.bams(bam.files, annotation.file, min.exon.count = 100,
  n.cores = defaultNCores(), min.umi.reads = 1)
}
\arguments{
\item{bam.files}{list of bam files (one per cell)}

\item{annotation.file}{gene annotation refFlat file}

\item{min.exon.count}{minimum number of molecules (across all cells) for the exon to be considered expressed}

\item{n.cores}{number of cores to use}

\item{min.umi.reads}{minimum number of read required per UMI/gene combination to be counted (defaults to 1)}
}
\value{
a list structure analogous to the return of read.smartseq2.bams(), counting molecules instead of reads.
}
\description{
Read in cell-specific bam files for STRT/C1
}
