% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_routines.R
\name{show.velocity.on.embedding.cor}
\alias{show.velocity.on.embedding.cor}
\title{Visualize RNA velocities on an existing embedding using correlation-based transition probability matrix within the kNN graph}
\usage{
show.velocity.on.embedding.cor(emb, vel, n = 100, cell.colors = NULL,
  corr.sigma = 0.05, show.grid.flow = FALSE, grid.n = 20,
  grid.sd = NULL, min.grid.cell.mass = 1, min.arrow.size = NULL,
  arrow.scale = 1, max.grid.arrow.length = NULL,
  fixed.arrow.length = FALSE, plot.grid.points = FALSE,
  scale = "log", nPcs = NA, arrow.lwd = 1, xlab = "", ylab = "",
  n.cores = defaultNCores(), do.par = T, show.cell = NULL,
  cell.border.alpha = 0.3, cc = NULL, return.details = FALSE,
  expression.scaling = FALSE, ...)
}
\arguments{
\item{emb}{embedding onto which to project the velocities; The dimensions of coordinates should be on the order of 10x10 for the default values to make sense.}

\item{vel}{velocity estimates (e.g. returned by gene.relative.velocity.estimates() )}

\item{n}{neighborhood size (default=100 cells)}

\item{cell.colors}{name vector of cell colors}

\item{corr.sigma}{sigma parameter used to translate velocity-(expression delta) correlation into a transition probability}

\item{show.grid.flow}{whether to show grid velocity summary}

\item{grid.n}{number of grid points along each axis}

\item{grid.sd}{standard deviation (in embedding coordinate space) used to determine the weighting of individual cells around each grid point}

\item{min.grid.cell.mass}{minimal cell "mass" (weighted number of cells) around each grid point required for it to show up}

\item{min.arrow.size}{minimal arrow size}

\item{arrow.scale}{arrow scale multiplier}

\item{max.grid.arrow.length}{minimal arrow size}

\item{fixed.arrow.length}{whether to use fixed arrow width (default=FALSE)}

\item{plot.grid.points}{whether to mark all grid points with dots (even if they don't have valid velocities)}

\item{scale}{velocity scale to use (default: 'log', other values: 'sqrt','rank','linear')}

\item{nPcs}{number of PCs to use for velocity regularization (default NA, turns off regularization)}

\item{arrow.lwd}{arrow width (under fixed.arrow.length=T)}

\item{xlab}{x axis label}

\item{ylab}{y axls label}

\item{n.cores}{number of cores to use}

\item{do.par}{whether to reset plotting parameters}

\item{show.cell}{whether to show detailed velocity estimates for a specified cell}

\item{cell.border.alpha}{transparency for the cell border}

\item{cc}{velocity-(exprssion delta) correlation matrix (can be passed back from previous results, as $cc) to save calculation time when replotting the same velocity estimates on the same embedding with different parameters}

\item{return.details}{whether to return detailed output (which can be passed to p1 app for visualization)}

\item{expression.scaling}{whether to scale the velocity length by the projection of velocity onto the expected expression change (based on the transition probability matrix)}

\item{...}{extra parameters passed to plot() function}
}
\value{
if return.details=F, returns invisible list containing transition probability matrix ($tp) and the velocity-(expression delta) correlation matrix ($cc). If return.details=T, returns a more extended list that can be passed as veloinfo to pagoda2::p2.make.pagoda1.app() for visualization
}
\description{
Visualize RNA velocities on an existing embedding using correlation-based transition probability matrix within the kNN graph
}
