% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_routines.R
\name{tSNE.velocity.plot}
\alias{tSNE.velocity.plot}
\title{Joint t-SNE visualization of the velocities}
\usage{
tSNE.velocity.plot(vel, cell.colors = NULL, scale = "log", plot.cols = 3,
  do.par = T, delta.norm = TRUE, nPcs = 15, norm.nPcs = nPcs * 10,
  perplexity = ncol(vel$current)/3, show.grid.flow = FALSE, grid.n = 20,
  grid.sd = NULL, min.grid.cell.mass = 1, pcount = 0.1, verbose = TRUE,
  min.arrow.median.ratio = 1/10, max.arrow.quantile = 0.9,
  arrow.scale = 1, arrow.lwd = 1, xlab = "", ylab = "",
  n.cores = defaultNCores(), size.norm = TRUE, ...)
}
\arguments{
\item{vel}{}

\item{cell.colors}{}

\item{scale}{}

\item{plot.cols}{}

\item{do.par}{}

\item{delta.norm}{}

\item{nPcs}{}

\item{norm.nPcs}{}

\item{perplexity}{}

\item{show.grid.flow}{}

\item{grid.n}{}

\item{grid.sd}{}

\item{min.grid.cell.mass}{}

\item{pcount}{}

\item{verbose}{}

\item{min.arrow.median.ratio}{}

\item{max.arrow.quantile}{}

\item{arrow.scale}{}

\item{arrow.lwd}{}

\item{xlab}{}

\item{ylab}{}

\item{n.cores}{}

\item{size.norm}{}

\item{...}{}
}
\description{
Joint t-SNE visualization of the velocities
}
