% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_routines.R
\name{tSNE.velocity.plot}
\alias{tSNE.velocity.plot}
\title{Joint t-SNE visualization of the velocities by joint t-SNE embedding of both current and extraploated cell positions}
\usage{
tSNE.velocity.plot(vel, cell.colors = NULL, scale = "log",
  do.par = T, delta.norm = TRUE, nPcs = 15, norm.nPcs = nPcs * 10,
  perplexity = ncol(vel$current)/3, show.grid.flow = FALSE,
  grid.n = 20, grid.sd = NULL, min.grid.cell.mass = 1,
  pcount = 0.1, verbose = TRUE, min.arrow.median.ratio = 1/10,
  max.arrow.quantile = 0.9, arrow.scale = 1, arrow.lwd = 1,
  xlab = "", ylab = "", n.cores = defaultNCores(),
  size.norm = TRUE, ...)
}
\arguments{
\item{vel}{velocity result}

\item{cell.colors}{named color vector for the cells}

\item{scale}{whether to rescale current/projected}

\item{do.par}{whether to reset par (default=T)}

\item{delta.norm}{whether to renormalize velocities following PCA projection}

\item{nPcs}{number of PCs onto which project the velocities}

\item{norm.nPcs}{number of PCs to use for velocity normalization}

\item{perplexity}{perplexity parameter to use in joint t-SNE calculation}

\item{show.grid.flow}{whether grid flow pattern should be drawn}

\item{grid.n}{number of grid points along each axis}

\item{grid.sd}{standard deviation of each grid point (used to determine the averaging radius for each grid point)}

\item{min.grid.cell.mass}{minimal number of cells around a grid point required for the grid point to show up}

\item{pcount}{pseudocount}

\item{verbose}{whether to show messages}

\item{min.arrow.median.ratio}{minimal ratio of arrow length (to the median arrow length) below which the arrows are not drawn (default=1/10)}

\item{max.arrow.quantile}{max arrow quantile that's used for arrow size calculation (default=0.9)}

\item{arrow.scale}{scaling factor for the arrows}

\item{arrow.lwd}{arrow line width}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{n.cores}{number of cores to use}

\item{size.norm}{whether to re-normalize current and projected cell sizes}

\item{...}{extra parameters are passed to plot() routine.}
}
\value{
invisible list containing embedding positions of current state (current.emb) and extrapolated states (projected.emb)
}
\description{
Joint t-SNE visualization of the velocities by joint t-SNE embedding of both current and extraploated cell positions
}
